backend/pyproject.toml
[tool.poetry]
name = "hc-doc-ai-wh-poc-backend"
version = "0.1.0"
description = "# Health Canada - DocAI Warehouse POC - Backend"
authors = ["Marcus Vinicius Monteiro de Souza <marcus.souza@improving.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = {extras = ["async"], version = "^3.0.0"}
flask-sqlalchemy = "^3.1.1"
sqlalchemy = "^2.0.25"
flask-migrate = "^4.0.5"
stream-unzip = "^0.0.89"
psycopg2-binary = "^2.9.9"
google-cloud-storage = "^2.14.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
python-dotenv = {extras = ["cli"], version = "^1.0.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
